airflow_ml_dags/requirements.txt
apache-airflow-providers-docker==1.2.0
apache-airflow-providers-ftp==1.1.0
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-sqlite==1.0.2
docker==3.7.3
click==7.1.2
numpy==1.19.1
pandas==1.2.0
PyYAML==5.3.1
scikit-learn==0.24.0
